requirements-dev.txt
WebOb==1.6.3
WebTest==2.0.23
beautifulsoup4==4.5.1
django-debug-toolbar==1.8
django-functest==1.0.1
django-webtest==1.8.0
furl==0.5.6
orderedmultidict==0.7.11
PyVirtualDisplay==0.2.1
selenium==2.53.6
sqlparse==0.2.3
waitress==1.0.1
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
fabric3==1.13.1.post1
paramiko==2.1.2
fabtools3==0.19.0
flake8==3.2.1
mccabe==0.5.2
pyflakes==1.3.0
pycodestyle==2.2.0
isort==4.2.5
ipython==5.3.0
unittest2==1.1.0
linecache2==1.0.0
traceback2==1.4.0
pyasn1==0.2.3
cryptography==1.9
idna==2.5
asn1crypto==0.22.0
cffi==1.10.0
pycparser==2.17
